+++ source/state_representation/test/tests/test_parameter.cpp
@@ -146,6 +146,11 @@ TYPED_TEST_P(ParameterTest, Construction) {
     EXPECT_FALSE(param.is_empty());
     EXPECT_TRUE(param);
     expect_values_equal(param.get_value(), std::get<0>(test_case));
+
+    param.initialize();
+    EXPECT_TRUE(param.is_empty());
+    EXPECT_FALSE(param);
+    expect_values_equal(param.get_value(), TypeParam());
   }
 }